Renamed winrm to winrmcommunicator
