+++ lib/vagrant-windows/communication/winrmcommunicator.rb
